requirements.txt
chromadb
sentence-transformers
google-genai
requests
beautifulsoup4
openai
streamlit
python-dotenv
rich
nltk
